edits

+++ src/net/fexcraft/app/fmt/ui/general/Scrollbar.java
@@ -75,7 +75,7 @@ public class Scrollbar extends Element {
 		
 		@Override
 		public void pullBy(int mx, int my){
-			scrollbar.scrolled += my * (scrollbar.element.getFullHeight() / scrollbar.height); if(scrollbar.scrolled < 0) scrollbar.scrolled = 0;
+			scrollbar.scrolled += my * (scrollbar.element.getFullHeight() / (float)scrollbar.height); if(scrollbar.scrolled < 0) scrollbar.scrolled = 0;
 			if(scrollbar.scrolled > scrollbar.element.getFullHeight()) scrollbar.scrolled = scrollbar.element.getFullHeight();
 		}
